requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Exchange
ccxt>=4.0.0
aiohttp>=3.9.0
websockets>=12.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# Database
aiosqlite>=0.19.0

# Web
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Utils
loguru>=0.7.0
schedule>=1.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
